+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add email protocol and SendGrid implementation
 - Make transit enable to use from cljs
 - Fix `qualified-name` to work on cljs
+- Upgrade Jetty9 #2
+- Add response types #3
 
 ## 0.2.1
 ### Added